Create Overloads To Allow Configuration via AppSettings
The library "serilog-settings-configuration" allows Serilog to read
from configuration provided by
"Microsoft.Extensions.Configuration". This integration has particular
limitations, and the existing method `AmazonCloudWatch` cannot be
used.

Three overloads have been created in order to support this:

- One allows the AWS access key, secret access key, and region to be
  read from the environment. Because of this, an implementation of
  `IAmazonCloudWatchLogs` with a default constructor (or a constructor
  with all optional arguments) may be provided via settings. If none
  is provided, an instance of `AmazonCloudWatchLogsClient` is
  created.
- Another takes in the region from settings, and reads the AWS access
  key and secret access key from the environment. This unconditionally
  creates an instance of `AmazonCloudWatchLogsClient`.
- Another takes in all values from settings. This also unconditionally
  creates an instance of `AmazonCloudWatchLogsClient`.

In all cases, the log group name is required, and the log event
renderer, the minimum log event level, the batch size limit, and the
period are optional, with default values. These default values are the
same as those used in `CloudWatchSinkOptions`.